refactoring to consolidate functions
