+++ src/lib.rs
@@ -1,0 +1,2 @@
+pub mod gtfs_rt;
+pub mod gtfs_static;
